Create Product Collection service and some changes to product controller class

+++ src/main/java/edu/icet/trendify/repository/product/ProductReviewRepository.java
@@ -14,7 +14,7 @@ public interface ProductReviewRepository extends JpaRepository<ProductReviewEnti
 
     List<ProductReviewEntity> findByCusId(Long cusId);
 
-    List<ProductReviewEntity> findByProductId(Long productId);
+    List<ProductReviewEntity> findByProductId(Integer productId);
 
     List<ProductReviewEntity> findByRating(Short rating);
 }
