+++ src/main/java/edu/icet/trendify/service/ProductService.java
@@ -15,7 +15,7 @@ public interface ProductService {
 
     Boolean addProductReview(ProductReviewDto reviewDto);
     Boolean updateProductReview(ProductReviewDto reviewDto);
-    List<ProductReviewDto> getProductReviewByProductId(Long id);
+    List<ProductReviewDto> getProductReviewByProductId(Integer id);
     List<ProductReviewDto> getProductReviewByCustomerId(Long id);
     List<ProductDto> getProductByRating(Short rating);
 }
